Fix Quitting Active Sessions and Infinite Ammo
Changed how active session data is handled to rectify issue with using qb target and quitting an active session. 
Received reports of players having infinite ammo in their weapon if it was the same as the test weapon, will need server owners to watch this issue and see if it continues to arise.